Create new PoolInfo interface
Currently PciDevice and its deviceType-specific derived versions
contain two types of information
- Low level device-specific information (such VF and PF names, PCI
  vendor, etc) used by the DeviceSelector implementations to filter
  devices
- High level API-centric information (mounts, env, deviceSpecs) used by
  the ResourcePool to respond to the API calls

This is fine as long as the Pool-specific information can derived solely
from the PCI information.

However, if a certain PCI device can be consumed in different ways
(like, for instance, a device compatible with DPDK and with DPDK-vDPA)
some configuration flag must be used to select which way the device
should be exposed. Unfortunately, the ResourceConfig is only available
when the ResorcePools are created (after the devices are).

In order to make that possible, split the two types of imformation into
two different interfaces (introducing PoolInfo). That way, the
ResourcePool-specific data can be generated taking into consideration
it's configuration.

The first benefit of this cleanup is that now the Rdma deviceSpec can be
appended to the rest of the device's deviceSpecs at pool-creation time
and only if the pool is properly configured, thus cleaning up the
resourcePool runtime code.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>