Fix a dependency error.
For some reason go mod was picking up an old version of client-go that
required an api (admissionregistration) that has been promoted from
alpha to beta.

I don't know if this is the best way to solve this issue but it does the
job for now.

TODO: Dig deeper and figure out if this is the right way

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>