pciDevice: Allow derived class to set the InfoProviders
The DeviceInfoProvider is currently being selected in the PciDevice base
class based on the driver being used.

That enforces a limitation on the device information that higher level
devices (pciNetDevices and accelDevices) can provide.

This patch allows the PciDevice constructor function to be given a
slice of DeviceInfoProvider implementations to be used. If empty, the
current default InformationProviders will be used for backwards
compatibility.

Since the Device Plugin supports returning a list of deviceSpecs and
mounts, we can append the information coming from each InfoProvider.
However, only one EnvVal is supported, so we only use the value returned
by the first InfoProvider.
This opens the door for a future improvement that adds support for
multiple EnvVars

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>